Refactor BPM renderers into provider packages
